Repos are added code inside this repo is pending

+++ src/DTO/IssueTicketRequest.java
@@ -8,7 +8,7 @@ import java.util.Date;
 
 public class IssueTicketRequest {
     private String vehicleNo;
-    private long gateId;
+    private int gateId;
     private String ownerName;
     private VehicleType vehicleType;
 
@@ -20,11 +20,11 @@ public class IssueTicketRequest {
         this.vehicleNo = vehicleNo;
     }
 
-    public long getGateId() {
+    public int getGateId() {
         return gateId;
     }
 
-    public void setGateId(long gateId) {
+    public void setGateId(int gateId) {
         this.gateId = gateId;
     }
